Fix a LGTM.com finding (false positive, but easier to just change it)

+++ payload.cc
@@ -191,7 +191,7 @@ struct token {
 static unsigned short *parse_portlist(const char *portlist, unsigned int *count) {
   uint32_t bitmap[65536 / 32];
   unsigned short *result;
-  unsigned short i;
+  unsigned int i;
   unsigned int p;
 
   memset(bitmap, 0, sizeof(bitmap));
